eliminate memory waste - simplify readability a little
